requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
mongomock
passlib[bcrypt]
python-jose[cryptography]
python-multipart
